+++ origami/collections/src/gold-elements/index.ts
@@ -1,1 +1,0 @@
-export * from './gold-elements.module';