+++ origami/collections/src/iron-elements/index.ts
@@ -1,1 +1,0 @@
-export * from './iron-elements.module';