+++ origami/collections/src/iron-elements/iron-input-elements.ts
@@ -1,7 +1,0 @@
-import { Directive } from '@angular/core';
-import { EmitChangesDirective } from '@codebakery/origami';
-
-@Directive({
-  selector: `iron-dropdown, iron-form, iron-input, iron-label`
-})
-export class IronInputElement extends EmitChangesDirective { }