+++ origami/collections/src/molecules/molecule-elements.ts
@@ -1,7 +1,0 @@
-import { Directive } from '@angular/core';
-import { EmitChangesDirective } from '@codebakery/origami';
-
-@Directive({
-  selector: `marked-element, prism-element`
-})
-export class MoleculeElement extends EmitChangesDirective { }