+++ origami/collections/src/paper-elements/paper-input-elements.ts
@@ -1,8 +1,0 @@
-import { Directive } from '@angular/core';
-import { EmitChangesDirective } from '@codebakery/origami';
-
-@Directive({
-  selector: `paper-checkbox, paper-dropdown-menu, paper-input, paper-listbox, paper-radio-button,
-    paper-radio-group, paper-slider, paper-toggle-button`
-})
-export class PaperInputElement extends EmitChangesDirective { }