+++ origami/src/polymer.module.ts
@@ -2,9 +2,7 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ɵDomSharedStylesHost as DomSharedStylesHost } from '@angular/platform-browser';
 
-import { EmitChangesDirective } from './events/emit-changes.directive';
 import { IronControlDirective } from './forms/iron-control.directive';
-import { CustomStyleService } from './style/custom-style.service';
 import { PolymerDomSharedStylesHost } from './style/shared-styles-host';
 import { PolymerTemplateDirective } from './templates/polymer-template.directive';
 
@@ -13,15 +11,10 @@ import { PolymerTemplateDirective } from './templates/polymer-template.directive
     FormsModule
   ],
   declarations: [
-    EmitChangesDirective,
     IronControlDirective,
     PolymerTemplateDirective
   ],
-  providers: [
-    CustomStyleService
-  ],
   exports: [
-    EmitChangesDirective,
     IronControlDirective,
     PolymerTemplateDirective
   ]