+++ origami/src/style/shared-styles-host.spec.ts
@@ -1,10 +1,11 @@
-// tslint:disable:max-classes-per-file
 import { Component, ElementRef, OnInit, ViewEncapsulation } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import {
   BrowserModule,
   ɵDomSharedStylesHost as DomSharedStylesHost
 } from '@angular/platform-browser';
+
+import 'polymer/polymer.html';
 
 import { PolymerDomSharedStylesHost } from './shared-styles-host';
 
@@ -17,7 +18,7 @@ import { PolymerDomSharedStylesHost } from './shared-styles-host';
       --red: red;
       --mixin: {
         font-size: 10px;
-      }
+      };
     }
     h1 {
       @apply --mixin;
@@ -28,34 +29,13 @@ import { PolymerDomSharedStylesHost } from './shared-styles-host';
 })
 class TestComponent { }
 
-@Component({
-  selector: 'native-component',
-  template: `<h1>Hello Native</h1>`,
-  styles: [
-    `
-    :host {
-      --blue: blue;
-      --mixin: {
-        font-size: 8px;
-      }
-    }
-    h1 {
-      @apply --mixin;
-      color: var(--blue);
-    }
-    `
-  ],
-  encapsulation: ViewEncapsulation.Native
-})
-class NativeComponent { }
-
 describe('PolymerDomSharedStylesHost', () => {
   let fixture: ComponentFixture<TestComponent>;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [BrowserModule],
-      declarations: [TestComponent, NativeComponent],
+      declarations: [TestComponent],
       providers: [{
         provide: DomSharedStylesHost, useClass: PolymerDomSharedStylesHost
       }]
@@ -64,43 +44,34 @@ describe('PolymerDomSharedStylesHost', () => {
     fixture = TestBed.createComponent(TestComponent);
   });
 
-  afterEach(() => {
-    Array.from(document.head.querySelectorAll('custom-style')).forEach(customStyle => {
-      document.head.removeChild(customStyle);
-      document.head.appendChild(customStyle.querySelector('style'));
-    });
-  });
-
-  it('should wrap <head> <style>s with <custom-style>', done => {
+  it('should add <style>s to ShadyCSS', done => {
     fixture.whenStable().then(() => {
       fixture.detectChanges();
-      expect(document.head.querySelectorAll('custom-style').length).toBeGreaterThan(0);
-      done();
-    });
-  });
-
-  if ('createShadowRoot' in HTMLElement.prototype) {
-    it('should wrap <style>s with <custom-style> in shadow DOM', done => {
-      fixture = TestBed.createComponent(NativeComponent);
-      fixture.whenStable().then(() => {
-        fixture.detectChanges();
-        expect(fixture.nativeElement.shadowRoot.querySelectorAll('custom-style').length)
-          .toBeGreaterThan(0);
-        done();
+      // Ignore scoped ShadyCSS-added style tags
+      Array.from(document.head.querySelectorAll('style:not([scope])')).forEach((style: any) => {
+        expect(style.__seenByShadyCSS).toBe(true);
       });
-    });
-  }
 
-  it('should ignore <style> elements with scope attribute from Polymer', done => {
-    const scopeStyle = document.createElement('style');
-    scopeStyle.setAttribute('scope', 'my-element');
-    document.head.appendChild(scopeStyle);
-    fixture.whenStable().then(() => {
-      fixture.detectChanges();
-      expect(document.head.querySelectorAll('custom-style').length).toBeGreaterThan(0);
-      expect(scopeStyle.parentElement).toBe(document.head);
-      document.head.removeChild(scopeStyle);
       done();
     });
+  });
+
+  it('should patch <body> to insert <style>s into the <head>', () => {
+    const div = document.createElement('div');
+    const style = document.createElement('style');
+    document.body.insertBefore(div, document.body.firstChild);
+    document.body.insertBefore(style, document.body.firstChild);
+    const divInBody = div.parentNode === document.body;
+    const styleInHead = style.parentNode === document.head;
+    if (div.parentElement) {
+      div.parentElement.removeChild(div);
+    }
+
+    if (style.parentElement) {
+      style.parentElement.removeChild(style);
+    }
+
+    expect(divInBody).toBe(true);
+    expect(styleInHead).toBe(true);
   });
 });