+++ origami/src/util/Polymer.spec.ts
@@ -1,8 +1,0 @@
-import { getPolymer } from './Polymer';
-
-describe('getPolymer()', () => {
-  it('should return window.Polymer', () => {
-    expect(window.Polymer).toBeDefined();
-    expect(getPolymer()).toBe(window.Polymer);
-  });
-});