+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.0.0-alpha.5"></a>
+# [2.0.0-alpha.5](https://github.com/hotforfeature/origami/compare/v2.0.0-alpha.4...v2.0.0-alpha.5) (2018-01-08)
+
+
+### Bug Fixes
+
+* **patch:** support Angular CLI 1.6 ([ec0e26d](https://github.com/hotforfeature/origami/commit/ec0e26d))
+
+
+
+<a name="2.0.0-alpha.4"></a>
+# [2.0.0-alpha.4](https://github.com/hotforfeature/origami/compare/v2.0.0-alpha.3...v2.0.0-alpha.4) (2017-11-15)
+
+This release fixes an NPM tag error.
+
+
+<a name="2.0.0-alpha.3"></a>
+# [2.0.0-alpha.3](https://github.com/hotforfeature/origami/compare/v2.0.0-alpha.2...v2.0.0-alpha.3) (2017-11-15)
+
+
+### Bug Fixes
+
+* **forms:** guard against ngOnDestroy called before ngOnInit ([46e53f0](https://github.com/hotforfeature/origami/commit/46e53f0))
+* **style:** stop adding ShadyCSS scoped styles as document-level styles (":not(.style-scope)" fix) ([d742059](https://github.com/hotforfeature/origami/commit/d742059))
+
+
+
+<a name="2.0.0-alpha.2"></a>
+# [2.0.0-alpha.2](https://github.com/hotforfeature/origami/compare/v2.0.0-alpha.1...v2.0.0-alpha.2) (2017-11-03)
+
+
+### Bug Fixes
+
+* compile with Angular 5 ([2634091](https://github.com/hotforfeature/origami/commit/2634091))
+
+
+
+<a name="2.0.0-alpha.1"></a>
+# [2.0.0-alpha.1](https://github.com/hotforfeature/origami/compare/v2.0.0-alpha.0...v2.0.0-alpha.1) (2017-11-03)
+
+
+### Bug Fixes
+
+* **patch:** update patch-cli for Angular CLI 1.5.0 ([34d370b](https://github.com/hotforfeature/origami/commit/34d370b))
+
+
+
 <a name="1.3.4"></a>
 ## [1.3.4](https://github.com/hotforfeature/origami/compare/v1.3.3...v1.3.4) (2017-11-20)
 
@@ -12,6 +59,7 @@ Fix metadata version mismatch.
 ## [1.3.3](https://github.com/hotforfeature/origami/compare/v1.3.2...v1.3.3) (2017-11-15)
 
 This release is to correct an NPM tag error.
+
 
 
 <a name="1.3.2"></a>